Move task definition from persistence layer to business logic layer (#1983)
* Move transfer tasks
* Move timer tasks
* Move visibility tasks
* Move replication tasks
* Update src & test accordingly
