+++ common/persistence/dataInterfaces.go
@@ -41,6 +41,7 @@ import (
 	persistencespb "go.temporal.io/server/api/persistence/v1"
 	"go.temporal.io/server/common/persistence/serialization"
 	"go.temporal.io/server/common/primitives"
+	"go.temporal.io/server/common/tasks"
 )
 
 // CreateWorkflowMode workflow creation mode
@@ -192,198 +193,11 @@ type (
 		*persistencespb.ReplicationTaskInfo
 	}
 
-	// Task is the generic interface for workflow tasks
-	Task interface {
-		GetType() enumsspb.TaskType
-		GetVersion() int64
-		SetVersion(version int64)
-		GetTaskID() int64
-		SetTaskID(id int64)
-		GetVisibilityTime() time.Time
-		SetVisibilityTime(timestamp time.Time)
-	}
-
 	// TaskQueueKey is the struct used to identity TaskQueues
 	TaskQueueKey struct {
 		NamespaceID string
 		Name        string
 		TaskType    enumspb.TaskQueueType
-	}
-
-	// ActivityTask identifies a transfer task for activity
-	ActivityTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		NamespaceID         string
-		TaskQueue           string
-		ScheduleID          int64
-		Version             int64
-	}
-
-	// WorkflowTask identifies a transfer task for workflow task
-	WorkflowTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		NamespaceID         string
-		TaskQueue           string
-		ScheduleID          int64
-		Version             int64
-	}
-
-	// ResetWorkflowTask identifies a transfer task to reset workflow
-	ResetWorkflowTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		Version             int64
-	}
-
-	// CloseExecutionTask identifies a transfer task for deletion of execution
-	CloseExecutionTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		Version             int64
-	}
-
-	// DeleteHistoryEventTask identifies a timer task for deletion of history events of completed execution.
-	DeleteHistoryEventTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		Version             int64
-	}
-
-	// WorkflowTaskTimeoutTask identifies a timeout task.
-	WorkflowTaskTimeoutTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		EventID             int64
-		ScheduleAttempt     int32
-		TimeoutType         enumspb.TimeoutType
-		Version             int64
-	}
-
-	// WorkflowTimeoutTask identifies a timeout task.
-	WorkflowTimeoutTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		Version             int64
-	}
-
-	// CancelExecutionTask identifies a transfer task for cancel of execution
-	CancelExecutionTask struct {
-		VisibilityTimestamp     time.Time
-		TaskID                  int64
-		TargetNamespaceID       string
-		TargetWorkflowID        string
-		TargetRunID             string
-		TargetChildWorkflowOnly bool
-		InitiatedID             int64
-		Version                 int64
-	}
-
-	// SignalExecutionTask identifies a transfer task for signal execution
-	SignalExecutionTask struct {
-		VisibilityTimestamp     time.Time
-		TaskID                  int64
-		TargetNamespaceID       string
-		TargetWorkflowID        string
-		TargetRunID             string
-		TargetChildWorkflowOnly bool
-		InitiatedID             int64
-		Version                 int64
-	}
-
-	// StartChildExecutionTask identifies a transfer task for starting child execution
-	StartChildExecutionTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		TargetNamespaceID   string
-		TargetWorkflowID    string
-		InitiatedID         int64
-		Version             int64
-	}
-
-	// ActivityTimeoutTask identifies a timeout task.
-	ActivityTimeoutTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		TimeoutType         enumspb.TimeoutType
-		EventID             int64
-		Attempt             int32
-		Version             int64
-	}
-
-	// UserTimerTask identifies a timeout task.
-	UserTimerTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		EventID             int64
-		Version             int64
-	}
-
-	// ActivityRetryTimerTask to schedule a retry task for activity
-	ActivityRetryTimerTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		EventID             int64
-		Version             int64
-		Attempt             int32
-	}
-
-	// WorkflowBackoffTimerTask to schedule first workflow task for retried workflow
-	WorkflowBackoffTimerTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		Version             int64
-		WorkflowBackoffType enumsspb.WorkflowBackoffType
-	}
-
-	// HistoryReplicationTask is the replication task created for shipping history replication events to other clusters
-	HistoryReplicationTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		FirstEventID        int64
-		NextEventID         int64
-		Version             int64
-		BranchToken         []byte
-		NewRunBranchToken   []byte
-	}
-
-	// StartExecutionVisibilityTask identifies a visibility task for start workflow execution.
-	StartExecutionVisibilityTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		Version             int64
-	}
-
-	// UpsertExecutionVisibilityTask identifies a visibility task for upsert workflow execution search attributes.
-	UpsertExecutionVisibilityTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		// this version is not used by task processing for validation,
-		// instead, the version is used by elastic search
-		Version int64
-	}
-
-	// CloseExecutionVisibilityTask identifies a visibility task for close workflow execution.
-	CloseExecutionVisibilityTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		Version             int64
-	}
-
-	// DeleteExecutionVisibilityTask identifies a visibility task for deletion of execution.
-	DeleteExecutionVisibilityTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		Version             int64
-	}
-
-	// SyncActivityTask is the replication task created for shipping activity info to other clusters
-	SyncActivityTask struct {
-		VisibilityTimestamp time.Time
-		TaskID              int64
-		Version             int64
-		ScheduledID         int64
 	}
 
 	// CreateShardRequest is used to create a shard in executions table
@@ -416,10 +230,10 @@ type (
 		WorkflowID  string
 		RunID       string
 
-		TransferTasks    []Task
-		TimerTasks       []Task
-		ReplicationTasks []Task
-		VisibilityTasks  []Task
+		TransferTasks    []tasks.Task
+		TimerTasks       []tasks.Task
+		ReplicationTasks []tasks.Task
+		VisibilityTasks  []tasks.Task
 	}
 
 	// CreateWorkflowExecutionRequest is used to write a new workflow execution
@@ -562,10 +376,10 @@ type (
 		NewBufferedEvents         []*historypb.HistoryEvent
 		ClearBufferedEvents       bool
 
-		TransferTasks    []Task
-		ReplicationTasks []Task
-		TimerTasks       []Task
-		VisibilityTasks  []Task
+		TransferTasks    []tasks.Task
+		ReplicationTasks []tasks.Task
+		TimerTasks       []tasks.Task
+		VisibilityTasks  []tasks.Task
 
 		// TODO deprecate Condition in favor of DBRecordVersion
 		Condition       int64
@@ -587,10 +401,10 @@ type (
 		SignalInfos         map[int64]*persistencespb.SignalInfo
 		SignalRequestedIDs  map[string]struct{}
 
-		TransferTasks    []Task
-		ReplicationTasks []Task
-		TimerTasks       []Task
-		VisibilityTasks  []Task
+		TransferTasks    []tasks.Task
+		ReplicationTasks []tasks.Task
+		TimerTasks       []tasks.Task
+		VisibilityTasks  []tasks.Task
 
 		// TODO deprecate Condition in favor of DBRecordVersion
 		Condition       int64
@@ -1379,709 +1193,9 @@ func (e *TransactionSizeLimitError) Error() string {
 	return e.Msg
 }
 
-// GetType returns the type of the activity task
-func (a *ActivityTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_TRANSFER_ACTIVITY_TASK
-}
-
-// GetVersion returns the version of the activity task
-func (a *ActivityTask) GetVersion() int64 {
-	return a.Version
-}
-
-// SetVersion returns the version of the activity task
-func (a *ActivityTask) SetVersion(version int64) {
-	a.Version = version
-}
-
-// GetTaskID returns the sequence ID of the activity task
-func (a *ActivityTask) GetTaskID() int64 {
-	return a.TaskID
-}
-
-// SetTaskID sets the sequence ID of the activity task
-func (a *ActivityTask) SetTaskID(id int64) {
-	a.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (a *ActivityTask) GetVisibilityTime() time.Time {
-	return a.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (a *ActivityTask) SetVisibilityTime(timestamp time.Time) {
-	a.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the workflow task
-func (d *WorkflowTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_TRANSFER_WORKFLOW_TASK
-}
-
-// GetVersion returns the version of the workflow task
-func (d *WorkflowTask) GetVersion() int64 {
-	return d.Version
-}
-
-// SetVersion returns the version of the workflow task
-func (d *WorkflowTask) SetVersion(version int64) {
-	d.Version = version
-}
-
-// GetTaskID returns the sequence ID of the workflow task.
-func (d *WorkflowTask) GetTaskID() int64 {
-	return d.TaskID
-}
-
-// SetTaskID sets the sequence ID of the workflow task
-func (d *WorkflowTask) SetTaskID(id int64) {
-	d.TaskID = id
-}
-
 // GetVisibilityTime get the visibility timestamp
 func (d *ReplicationTaskInfoWrapper) GetVisibilityTime() *time.Time {
 	return &time.Time{}
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (d *WorkflowTask) GetVisibilityTime() time.Time {
-	return d.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (d *WorkflowTask) SetVisibilityTime(timestamp time.Time) {
-	d.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the ResetWorkflowTask
-func (a *ResetWorkflowTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_TRANSFER_RESET_WORKFLOW
-}
-
-// GetVersion returns the version of the ResetWorkflowTask
-func (a *ResetWorkflowTask) GetVersion() int64 {
-	return a.Version
-}
-
-// SetVersion returns the version of the ResetWorkflowTask
-func (a *ResetWorkflowTask) SetVersion(version int64) {
-	a.Version = version
-}
-
-// GetTaskID returns the sequence ID of the ResetWorkflowTask
-func (a *ResetWorkflowTask) GetTaskID() int64 {
-	return a.TaskID
-}
-
-// SetTaskID sets the sequence ID of the ResetWorkflowTask
-func (a *ResetWorkflowTask) SetTaskID(id int64) {
-	a.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (a *ResetWorkflowTask) GetVisibilityTime() time.Time {
-	return a.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (a *ResetWorkflowTask) SetVisibilityTime(timestamp time.Time) {
-	a.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the close execution task
-func (a *CloseExecutionTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_TRANSFER_CLOSE_EXECUTION
-}
-
-// GetVersion returns the version of the close execution task
-func (a *CloseExecutionTask) GetVersion() int64 {
-	return a.Version
-}
-
-// SetVersion returns the version of the close execution task
-func (a *CloseExecutionTask) SetVersion(version int64) {
-	a.Version = version
-}
-
-// GetTaskID returns the sequence ID of the close execution task
-func (a *CloseExecutionTask) GetTaskID() int64 {
-	return a.TaskID
-}
-
-// SetTaskID sets the sequence ID of the close execution task
-func (a *CloseExecutionTask) SetTaskID(id int64) {
-	a.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (a *CloseExecutionTask) GetVisibilityTime() time.Time {
-	return a.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (a *CloseExecutionTask) SetVisibilityTime(timestamp time.Time) {
-	a.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the delete execution task
-func (a *DeleteHistoryEventTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_DELETE_HISTORY_EVENT
-}
-
-// GetVersion returns the version of the delete execution task
-func (a *DeleteHistoryEventTask) GetVersion() int64 {
-	return a.Version
-}
-
-// SetVersion returns the version of the delete execution task
-func (a *DeleteHistoryEventTask) SetVersion(version int64) {
-	a.Version = version
-}
-
-// GetTaskID returns the sequence ID of the delete execution task
-func (a *DeleteHistoryEventTask) GetTaskID() int64 {
-	return a.TaskID
-}
-
-// SetTaskID sets the sequence ID of the delete execution task
-func (a *DeleteHistoryEventTask) SetTaskID(id int64) {
-	a.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (a *DeleteHistoryEventTask) GetVisibilityTime() time.Time {
-	return a.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (a *DeleteHistoryEventTask) SetVisibilityTime(timestamp time.Time) {
-	a.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the timer task
-func (d *WorkflowTaskTimeoutTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_WORKFLOW_TASK_TIMEOUT
-}
-
-// GetVersion returns the version of the timer task
-func (d *WorkflowTaskTimeoutTask) GetVersion() int64 {
-	return d.Version
-}
-
-// SetVersion returns the version of the timer task
-func (d *WorkflowTaskTimeoutTask) SetVersion(version int64) {
-	d.Version = version
-}
-
-// GetTaskID returns the sequence ID.
-func (d *WorkflowTaskTimeoutTask) GetTaskID() int64 {
-	return d.TaskID
-}
-
-// SetTaskID sets the sequence ID.
-func (d *WorkflowTaskTimeoutTask) SetTaskID(id int64) {
-	d.TaskID = id
-}
-
-// GetVisibilityTime gets the visibility time stamp
-func (d *WorkflowTaskTimeoutTask) GetVisibilityTime() time.Time {
-	return d.VisibilityTimestamp
-}
-
-// SetVisibilityTime gets the visibility time stamp
-func (d *WorkflowTaskTimeoutTask) SetVisibilityTime(t time.Time) {
-	d.VisibilityTimestamp = t
-}
-
-// GetType returns the type of the timer task
-func (a *ActivityTimeoutTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_ACTIVITY_TIMEOUT
-}
-
-// GetVersion returns the version of the timer task
-func (a *ActivityTimeoutTask) GetVersion() int64 {
-	return a.Version
-}
-
-// SetVersion returns the version of the timer task
-func (a *ActivityTimeoutTask) SetVersion(version int64) {
-	a.Version = version
-}
-
-// GetTaskID returns the sequence ID.
-func (a *ActivityTimeoutTask) GetTaskID() int64 {
-	return a.TaskID
-}
-
-// SetTaskID sets the sequence ID.
-func (a *ActivityTimeoutTask) SetTaskID(id int64) {
-	a.TaskID = id
-}
-
-// GetVisibilityTime gets the visibility time stamp
-func (a *ActivityTimeoutTask) GetVisibilityTime() time.Time {
-	return a.VisibilityTimestamp
-}
-
-// SetVisibilityTime gets the visibility time stamp
-func (a *ActivityTimeoutTask) SetVisibilityTime(t time.Time) {
-	a.VisibilityTimestamp = t
-}
-
-// GetType returns the type of the timer task
-func (u *UserTimerTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_USER_TIMER
-}
-
-// GetVersion returns the version of the timer task
-func (u *UserTimerTask) GetVersion() int64 {
-	return u.Version
-}
-
-// SetVersion returns the version of the timer task
-func (u *UserTimerTask) SetVersion(version int64) {
-	u.Version = version
-}
-
-// GetTaskID returns the sequence ID of the timer task.
-func (u *UserTimerTask) GetTaskID() int64 {
-	return u.TaskID
-}
-
-// SetTaskID sets the sequence ID of the timer task.
-func (u *UserTimerTask) SetTaskID(id int64) {
-	u.TaskID = id
-}
-
-// GetVisibilityTime gets the visibility time stamp
-func (u *UserTimerTask) GetVisibilityTime() time.Time {
-	return u.VisibilityTimestamp
-}
-
-// SetVisibilityTime gets the visibility time stamp
-func (u *UserTimerTask) SetVisibilityTime(t time.Time) {
-	u.VisibilityTimestamp = t
-}
-
-// GetType returns the type of the retry timer task
-func (r *ActivityRetryTimerTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_ACTIVITY_RETRY_TIMER
-}
-
-// GetVersion returns the version of the retry timer task
-func (r *ActivityRetryTimerTask) GetVersion() int64 {
-	return r.Version
-}
-
-// SetVersion returns the version of the retry timer task
-func (r *ActivityRetryTimerTask) SetVersion(version int64) {
-	r.Version = version
-}
-
-// GetTaskID returns the sequence ID.
-func (r *ActivityRetryTimerTask) GetTaskID() int64 {
-	return r.TaskID
-}
-
-// SetTaskID sets the sequence ID.
-func (r *ActivityRetryTimerTask) SetTaskID(id int64) {
-	r.TaskID = id
-}
-
-// GetVisibilityTime gets the visibility time stamp
-func (r *ActivityRetryTimerTask) GetVisibilityTime() time.Time {
-	return r.VisibilityTimestamp
-}
-
-// SetVisibilityTime gets the visibility time stamp
-func (r *ActivityRetryTimerTask) SetVisibilityTime(t time.Time) {
-	r.VisibilityTimestamp = t
-}
-
-// GetType returns the type of the retry timer task
-func (r *WorkflowBackoffTimerTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_WORKFLOW_BACKOFF_TIMER
-}
-
-// GetVersion returns the version of the retry timer task
-func (r *WorkflowBackoffTimerTask) GetVersion() int64 {
-	return r.Version
-}
-
-// SetVersion returns the version of the retry timer task
-func (r *WorkflowBackoffTimerTask) SetVersion(version int64) {
-	r.Version = version
-}
-
-// GetTaskID returns the sequence ID.
-func (r *WorkflowBackoffTimerTask) GetTaskID() int64 {
-	return r.TaskID
-}
-
-// SetTaskID sets the sequence ID.
-func (r *WorkflowBackoffTimerTask) SetTaskID(id int64) {
-	r.TaskID = id
-}
-
-// GetVisibilityTime gets the visibility time stamp
-func (r *WorkflowBackoffTimerTask) GetVisibilityTime() time.Time {
-	return r.VisibilityTimestamp
-}
-
-// SetVisibilityTime gets the visibility time stamp
-func (r *WorkflowBackoffTimerTask) SetVisibilityTime(t time.Time) {
-	r.VisibilityTimestamp = t
-}
-
-// GetType returns the type of the timeout task.
-func (u *WorkflowTimeoutTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_WORKFLOW_RUN_TIMEOUT
-}
-
-// GetVersion returns the version of the timeout task
-func (u *WorkflowTimeoutTask) GetVersion() int64 {
-	return u.Version
-}
-
-// SetVersion returns the version of the timeout task
-func (u *WorkflowTimeoutTask) SetVersion(version int64) {
-	u.Version = version
-}
-
-// GetTaskID returns the sequence ID of the cancel transfer task.
-func (u *WorkflowTimeoutTask) GetTaskID() int64 {
-	return u.TaskID
-}
-
-// SetTaskID sets the sequence ID of the cancel transfer task.
-func (u *WorkflowTimeoutTask) SetTaskID(id int64) {
-	u.TaskID = id
-}
-
-// GetVisibilityTime gets the visibility time stamp
-func (u *WorkflowTimeoutTask) GetVisibilityTime() time.Time {
-	return u.VisibilityTimestamp
-}
-
-// SetVisibilityTime gets the visibility time stamp
-func (u *WorkflowTimeoutTask) SetVisibilityTime(t time.Time) {
-	u.VisibilityTimestamp = t
-}
-
-// GetType returns the type of the cancel transfer task
-func (u *CancelExecutionTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_TRANSFER_CANCEL_EXECUTION
-}
-
-// GetVersion returns the version of the cancel transfer task
-func (u *CancelExecutionTask) GetVersion() int64 {
-	return u.Version
-}
-
-// SetVersion returns the version of the cancel transfer task
-func (u *CancelExecutionTask) SetVersion(version int64) {
-	u.Version = version
-}
-
-// GetTaskID returns the sequence ID of the cancel transfer task.
-func (u *CancelExecutionTask) GetTaskID() int64 {
-	return u.TaskID
-}
-
-// SetTaskID sets the sequence ID of the cancel transfer task.
-func (u *CancelExecutionTask) SetTaskID(id int64) {
-	u.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (u *CancelExecutionTask) GetVisibilityTime() time.Time {
-	return u.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (u *CancelExecutionTask) SetVisibilityTime(timestamp time.Time) {
-	u.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the signal transfer task
-func (u *SignalExecutionTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_TRANSFER_SIGNAL_EXECUTION
-}
-
-// GetVersion returns the version of the signal transfer task
-func (u *SignalExecutionTask) GetVersion() int64 {
-	return u.Version
-}
-
-// SetVersion returns the version of the signal transfer task
-func (u *SignalExecutionTask) SetVersion(version int64) {
-	u.Version = version
-}
-
-// GetTaskID returns the sequence ID of the signal transfer task.
-func (u *SignalExecutionTask) GetTaskID() int64 {
-	return u.TaskID
-}
-
-// SetTaskID sets the sequence ID of the signal transfer task.
-func (u *SignalExecutionTask) SetTaskID(id int64) {
-	u.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (u *SignalExecutionTask) GetVisibilityTime() time.Time {
-	return u.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (u *SignalExecutionTask) SetVisibilityTime(timestamp time.Time) {
-	u.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the start child transfer task
-func (u *StartChildExecutionTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_TRANSFER_START_CHILD_EXECUTION
-}
-
-// GetVersion returns the version of the start child transfer task
-func (u *StartChildExecutionTask) GetVersion() int64 {
-	return u.Version
-}
-
-// SetVersion returns the version of the start child transfer task
-func (u *StartChildExecutionTask) SetVersion(version int64) {
-	u.Version = version
-}
-
-// GetTaskID returns the sequence ID of the start child transfer task
-func (u *StartChildExecutionTask) GetTaskID() int64 {
-	return u.TaskID
-}
-
-// SetTaskID sets the sequence ID of the start child transfer task
-func (u *StartChildExecutionTask) SetTaskID(id int64) {
-	u.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (u *StartChildExecutionTask) GetVisibilityTime() time.Time {
-	return u.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (u *StartChildExecutionTask) SetVisibilityTime(timestamp time.Time) {
-	u.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the history replication task
-func (a *HistoryReplicationTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_REPLICATION_HISTORY
-}
-
-// GetVersion returns the version of the history replication task
-func (a *HistoryReplicationTask) GetVersion() int64 {
-	return a.Version
-}
-
-// SetVersion returns the version of the history replication task
-func (a *HistoryReplicationTask) SetVersion(version int64) {
-	a.Version = version
-}
-
-// GetTaskID returns the sequence ID of the history replication task
-func (a *HistoryReplicationTask) GetTaskID() int64 {
-	return a.TaskID
-}
-
-// SetTaskID sets the sequence ID of the history replication task
-func (a *HistoryReplicationTask) SetTaskID(id int64) {
-	a.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (a *HistoryReplicationTask) GetVisibilityTime() time.Time {
-	return a.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (a *HistoryReplicationTask) SetVisibilityTime(timestamp time.Time) {
-	a.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the upsert search attributes transfer task
-func (t *StartExecutionVisibilityTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_VISIBILITY_START_EXECUTION
-}
-
-// GetVersion returns the version of the upsert search attributes transfer task
-func (t *StartExecutionVisibilityTask) GetVersion() int64 {
-	return t.Version
-}
-
-// SetVersion returns the version of the upsert search attributes transfer task
-func (t *StartExecutionVisibilityTask) SetVersion(version int64) {
-	t.Version = version
-}
-
-// GetTaskID returns the sequence ID of the signal transfer task.
-func (t *StartExecutionVisibilityTask) GetTaskID() int64 {
-	return t.TaskID
-}
-
-// SetTaskID sets the sequence ID of the signal transfer task.
-func (t *StartExecutionVisibilityTask) SetTaskID(id int64) {
-	t.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (t *StartExecutionVisibilityTask) GetVisibilityTime() time.Time {
-	return t.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (t *StartExecutionVisibilityTask) SetVisibilityTime(timestamp time.Time) {
-	t.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the upsert search attributes transfer task
-func (t *UpsertExecutionVisibilityTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_VISIBILITY_UPSERT_EXECUTION
-}
-
-// GetVersion returns the version of the upsert search attributes transfer task
-func (t *UpsertExecutionVisibilityTask) GetVersion() int64 {
-	return t.Version
-}
-
-// SetVersion returns the version of the upsert search attributes transfer task
-func (t *UpsertExecutionVisibilityTask) SetVersion(version int64) {
-	t.Version = version
-}
-
-// GetTaskID returns the sequence ID of the signal transfer task.
-func (t *UpsertExecutionVisibilityTask) GetTaskID() int64 {
-	return t.TaskID
-}
-
-// SetTaskID sets the sequence ID of the signal transfer task.
-func (t *UpsertExecutionVisibilityTask) SetTaskID(id int64) {
-	t.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (t *UpsertExecutionVisibilityTask) GetVisibilityTime() time.Time {
-	return t.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (t *UpsertExecutionVisibilityTask) SetVisibilityTime(timestamp time.Time) {
-	t.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the upsert search attributes transfer task
-func (t *CloseExecutionVisibilityTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_VISIBILITY_CLOSE_EXECUTION
-}
-
-// GetVersion returns the version of the upsert search attributes transfer task
-func (t *CloseExecutionVisibilityTask) GetVersion() int64 {
-	return t.Version
-}
-
-// SetVersion returns the version of the upsert search attributes transfer task
-func (t *CloseExecutionVisibilityTask) SetVersion(version int64) {
-	t.Version = version
-}
-
-// GetTaskID returns the sequence ID of the signal transfer task.
-func (t *CloseExecutionVisibilityTask) GetTaskID() int64 {
-	return t.TaskID
-}
-
-// SetTaskID sets the sequence ID of the signal transfer task.
-func (t *CloseExecutionVisibilityTask) SetTaskID(id int64) {
-	t.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (t *CloseExecutionVisibilityTask) GetVisibilityTime() time.Time {
-	return t.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (t *CloseExecutionVisibilityTask) SetVisibilityTime(timestamp time.Time) {
-	t.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the upsert search attributes transfer task
-func (t *DeleteExecutionVisibilityTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_VISIBILITY_DELETE_EXECUTION
-}
-
-// GetVersion returns the version of the upsert search attributes transfer task
-func (t *DeleteExecutionVisibilityTask) GetVersion() int64 {
-	return t.Version
-}
-
-// SetVersion returns the version of the upsert search attributes transfer task
-func (t *DeleteExecutionVisibilityTask) SetVersion(version int64) {
-	t.Version = version
-}
-
-// GetTaskID returns the sequence ID of the signal transfer task.
-func (t *DeleteExecutionVisibilityTask) GetTaskID() int64 {
-	return t.TaskID
-}
-
-// SetTaskID sets the sequence ID of the signal transfer task.
-func (t *DeleteExecutionVisibilityTask) SetTaskID(id int64) {
-	t.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (t *DeleteExecutionVisibilityTask) GetVisibilityTime() time.Time {
-	return t.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (t *DeleteExecutionVisibilityTask) SetVisibilityTime(timestamp time.Time) {
-	t.VisibilityTimestamp = timestamp
-}
-
-// GetType returns the type of the history replication task
-func (a *SyncActivityTask) GetType() enumsspb.TaskType {
-	return enumsspb.TASK_TYPE_REPLICATION_SYNC_ACTIVITY
-}
-
-// GetVersion returns the version of the history replication task
-func (a *SyncActivityTask) GetVersion() int64 {
-	return a.Version
-}
-
-// SetVersion returns the version of the history replication task
-func (a *SyncActivityTask) SetVersion(version int64) {
-	a.Version = version
-}
-
-// GetTaskID returns the sequence ID of the history replication task
-func (a *SyncActivityTask) GetTaskID() int64 {
-	return a.TaskID
-}
-
-// SetTaskID sets the sequence ID of the history replication task
-func (a *SyncActivityTask) SetTaskID(id int64) {
-	a.TaskID = id
-}
-
-// GetVisibilityTime get the visibility timestamp
-func (a *SyncActivityTask) GetVisibilityTime() time.Time {
-	return a.VisibilityTimestamp
-}
-
-// SetVisibilityTime set the visibility timestamp
-func (a *SyncActivityTask) SetVisibilityTime(timestamp time.Time) {
-	a.VisibilityTimestamp = timestamp
 }
 
 // UnixMilliseconds returns t as a Unix time, the number of milliseconds elapsed since January 1, 1970 UTC.
