+++ common/persistence/dataInterfaces_mock.go
@@ -30,126 +30,9 @@ package persistence
 
 import (
 	reflect "reflect"
-	time "time"
 
 	gomock "github.com/golang/mock/gomock"
-	v1 "go.temporal.io/server/api/enums/v1"
 )
-
-// MockTask is a mock of Task interface.
-type MockTask struct {
-	ctrl     *gomock.Controller
-	recorder *MockTaskMockRecorder
-}
-
-// MockTaskMockRecorder is the mock recorder for MockTask.
-type MockTaskMockRecorder struct {
-	mock *MockTask
-}
-
-// NewMockTask creates a new mock instance.
-func NewMockTask(ctrl *gomock.Controller) *MockTask {
-	mock := &MockTask{ctrl: ctrl}
-	mock.recorder = &MockTaskMockRecorder{mock}
-	return mock
-}
-
-// EXPECT returns an object that allows the caller to indicate expected use.
-func (m *MockTask) EXPECT() *MockTaskMockRecorder {
-	return m.recorder
-}
-
-// GetTaskID mocks base method.
-func (m *MockTask) GetTaskID() int64 {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetTaskID")
-	ret0, _ := ret[0].(int64)
-	return ret0
-}
-
-// GetTaskID indicates an expected call of GetTaskID.
-func (mr *MockTaskMockRecorder) GetTaskID() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetTaskID", reflect.TypeOf((*MockTask)(nil).GetTaskID))
-}
-
-// GetType mocks base method.
-func (m *MockTask) GetType() v1.TaskType {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetType")
-	ret0, _ := ret[0].(v1.TaskType)
-	return ret0
-}
-
-// GetType indicates an expected call of GetType.
-func (mr *MockTaskMockRecorder) GetType() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetType", reflect.TypeOf((*MockTask)(nil).GetType))
-}
-
-// GetVersion mocks base method.
-func (m *MockTask) GetVersion() int64 {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetVersion")
-	ret0, _ := ret[0].(int64)
-	return ret0
-}
-
-// GetVersion indicates an expected call of GetVersion.
-func (mr *MockTaskMockRecorder) GetVersion() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetVersion", reflect.TypeOf((*MockTask)(nil).GetVersion))
-}
-
-// GetVisibilityTime mocks base method.
-func (m *MockTask) GetVisibilityTime() time.Time {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetVisibilityTime")
-	ret0, _ := ret[0].(time.Time)
-	return ret0
-}
-
-// GetVisibilityTime indicates an expected call of GetVisibilityTime.
-func (mr *MockTaskMockRecorder) GetVisibilityTime() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetVisibilityTime", reflect.TypeOf((*MockTask)(nil).GetVisibilityTime))
-}
-
-// SetTaskID mocks base method.
-func (m *MockTask) SetTaskID(id int64) {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "SetTaskID", id)
-}
-
-// SetTaskID indicates an expected call of SetTaskID.
-func (mr *MockTaskMockRecorder) SetTaskID(id interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SetTaskID", reflect.TypeOf((*MockTask)(nil).SetTaskID), id)
-}
-
-// SetVersion mocks base method.
-func (m *MockTask) SetVersion(version int64) {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "SetVersion", version)
-}
-
-// SetVersion indicates an expected call of SetVersion.
-func (mr *MockTaskMockRecorder) SetVersion(version interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SetVersion", reflect.TypeOf((*MockTask)(nil).SetVersion), version)
-}
-
-// SetVisibilityTime mocks base method.
-func (m *MockTask) SetVisibilityTime(timestamp time.Time) {
-	m.ctrl.T.Helper()
-	m.ctrl.Call(m, "SetVisibilityTime", timestamp)
-}
-
-// SetVisibilityTime indicates an expected call of SetVisibilityTime.
-func (mr *MockTaskMockRecorder) SetVisibilityTime(timestamp interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SetVisibilityTime", reflect.TypeOf((*MockTask)(nil).SetVisibilityTime), timestamp)
-}
 
 // MockCloseable is a mock of Closeable interface.
 type MockCloseable struct {
