+++ common/tasks/activity_retry_timer.go
@@ -22,52 +22,49 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package workflow
+package tasks
 
 import (
 	"time"
 
-	persistencespb "go.temporal.io/server/api/persistence/v1"
-	"go.temporal.io/server/common/tasks"
+	enumsspb "go.temporal.io/server/api/enums/v1"
 )
 
-// NOTE: do not use make(type, len(input))
-// since this will assume initial length being len(inputs)
-// always use make(type, 0, len(input))
-func convertSyncActivityInfos(
-	activityInfos map[int64]*persistencespb.ActivityInfo,
-	inputs map[int64]struct{},
-) []tasks.Task {
-	outputs := make([]tasks.Task, 0, len(inputs))
-	for item := range inputs {
-		activityInfo, ok := activityInfos[item]
-		if ok {
-			outputs = append(outputs, &tasks.SyncActivityTask{
-				Version:     activityInfo.Version,
-				ScheduledID: activityInfo.ScheduleId,
-			})
-		}
+type (
+	ActivityRetryTimerTask struct {
+		VisibilityTimestamp time.Time
+		TaskID              int64
+		EventID             int64
+		Version             int64
+		Attempt             int32
 	}
-	return outputs
+)
+
+// GetType returns the type of the retry timer task
+func (r *ActivityRetryTimerTask) GetType() enumsspb.TaskType {
+	return enumsspb.TASK_TYPE_ACTIVITY_RETRY_TIMER
 }
 
-func setTaskInfo(
-	version int64,
-	timestamp time.Time,
-	transferTasks []tasks.Task,
-	timerTasks []tasks.Task,
-	visibilityTasks []tasks.Task,
-) {
-	// set both the task version, as well as the Timestamp on the transfer tasks
-	for _, task := range transferTasks {
-		task.SetVersion(version)
-		task.SetVisibilityTime(timestamp)
-	}
-	for _, task := range timerTasks {
-		task.SetVersion(version)
-	}
-	for _, task := range visibilityTasks {
-		task.SetVersion(version)
-		task.SetVisibilityTime(timestamp)
-	}
+func (r *ActivityRetryTimerTask) GetVersion() int64 {
+	return r.Version
+}
+
+func (r *ActivityRetryTimerTask) SetVersion(version int64) {
+	r.Version = version
+}
+
+func (r *ActivityRetryTimerTask) GetTaskID() int64 {
+	return r.TaskID
+}
+
+func (r *ActivityRetryTimerTask) SetTaskID(id int64) {
+	r.TaskID = id
+}
+
+func (r *ActivityRetryTimerTask) GetVisibilityTime() time.Time {
+	return r.VisibilityTimestamp
+}
+
+func (r *ActivityRetryTimerTask) SetVisibilityTime(t time.Time) {
+	r.VisibilityTimestamp = t
 }
