+++ common/tasks/activity_task_timer.go
@@ -22,52 +22,51 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package workflow
+package tasks
 
 import (
 	"time"
 
-	persistencespb "go.temporal.io/server/api/persistence/v1"
-	"go.temporal.io/server/common/tasks"
+	enumspb "go.temporal.io/api/enums/v1"
+
+	enumsspb "go.temporal.io/server/api/enums/v1"
 )
 
-// NOTE: do not use make(type, len(input))
-// since this will assume initial length being len(inputs)
-// always use make(type, 0, len(input))
-func convertSyncActivityInfos(
-	activityInfos map[int64]*persistencespb.ActivityInfo,
-	inputs map[int64]struct{},
-) []tasks.Task {
-	outputs := make([]tasks.Task, 0, len(inputs))
-	for item := range inputs {
-		activityInfo, ok := activityInfos[item]
-		if ok {
-			outputs = append(outputs, &tasks.SyncActivityTask{
-				Version:     activityInfo.Version,
-				ScheduledID: activityInfo.ScheduleId,
-			})
-		}
+type (
+	ActivityTimeoutTask struct {
+		VisibilityTimestamp time.Time
+		TaskID              int64
+		TimeoutType         enumspb.TimeoutType
+		EventID             int64
+		Attempt             int32
+		Version             int64
 	}
-	return outputs
+)
+
+func (a *ActivityTimeoutTask) GetType() enumsspb.TaskType {
+	return enumsspb.TASK_TYPE_ACTIVITY_TIMEOUT
 }
 
-func setTaskInfo(
-	version int64,
-	timestamp time.Time,
-	transferTasks []tasks.Task,
-	timerTasks []tasks.Task,
-	visibilityTasks []tasks.Task,
-) {
-	// set both the task version, as well as the Timestamp on the transfer tasks
-	for _, task := range transferTasks {
-		task.SetVersion(version)
-		task.SetVisibilityTime(timestamp)
-	}
-	for _, task := range timerTasks {
-		task.SetVersion(version)
-	}
-	for _, task := range visibilityTasks {
-		task.SetVersion(version)
-		task.SetVisibilityTime(timestamp)
-	}
+func (a *ActivityTimeoutTask) GetVersion() int64 {
+	return a.Version
+}
+
+func (a *ActivityTimeoutTask) SetVersion(version int64) {
+	a.Version = version
+}
+
+func (a *ActivityTimeoutTask) GetTaskID() int64 {
+	return a.TaskID
+}
+
+func (a *ActivityTimeoutTask) SetTaskID(id int64) {
+	a.TaskID = id
+}
+
+func (a *ActivityTimeoutTask) GetVisibilityTime() time.Time {
+	return a.VisibilityTimestamp
+}
+
+func (a *ActivityTimeoutTask) SetVisibilityTime(t time.Time) {
+	a.VisibilityTimestamp = t
 }
