+++ common/tasks/start_visibility_task.go
@@ -22,52 +22,46 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package workflow
+package tasks
 
 import (
 	"time"
 
-	persistencespb "go.temporal.io/server/api/persistence/v1"
-	"go.temporal.io/server/common/tasks"
+	enumsspb "go.temporal.io/server/api/enums/v1"
 )
 
-// NOTE: do not use make(type, len(input))
-// since this will assume initial length being len(inputs)
-// always use make(type, 0, len(input))
-func convertSyncActivityInfos(
-	activityInfos map[int64]*persistencespb.ActivityInfo,
-	inputs map[int64]struct{},
-) []tasks.Task {
-	outputs := make([]tasks.Task, 0, len(inputs))
-	for item := range inputs {
-		activityInfo, ok := activityInfos[item]
-		if ok {
-			outputs = append(outputs, &tasks.SyncActivityTask{
-				Version:     activityInfo.Version,
-				ScheduledID: activityInfo.ScheduleId,
-			})
-		}
+type (
+	StartExecutionVisibilityTask struct {
+		VisibilityTimestamp time.Time
+		TaskID              int64
+		Version             int64
 	}
-	return outputs
+)
+
+func (t *StartExecutionVisibilityTask) GetType() enumsspb.TaskType {
+	return enumsspb.TASK_TYPE_VISIBILITY_START_EXECUTION
 }
 
-func setTaskInfo(
-	version int64,
-	timestamp time.Time,
-	transferTasks []tasks.Task,
-	timerTasks []tasks.Task,
-	visibilityTasks []tasks.Task,
-) {
-	// set both the task version, as well as the Timestamp on the transfer tasks
-	for _, task := range transferTasks {
-		task.SetVersion(version)
-		task.SetVisibilityTime(timestamp)
-	}
-	for _, task := range timerTasks {
-		task.SetVersion(version)
-	}
-	for _, task := range visibilityTasks {
-		task.SetVersion(version)
-		task.SetVisibilityTime(timestamp)
-	}
+func (t *StartExecutionVisibilityTask) GetVersion() int64 {
+	return t.Version
+}
+
+func (t *StartExecutionVisibilityTask) SetVersion(version int64) {
+	t.Version = version
+}
+
+func (t *StartExecutionVisibilityTask) GetTaskID() int64 {
+	return t.TaskID
+}
+
+func (t *StartExecutionVisibilityTask) SetTaskID(id int64) {
+	t.TaskID = id
+}
+
+func (t *StartExecutionVisibilityTask) GetVisibilityTime() time.Time {
+	return t.VisibilityTimestamp
+}
+
+func (t *StartExecutionVisibilityTask) SetVisibilityTime(timestamp time.Time) {
+	t.VisibilityTimestamp = timestamp
 }
