+++ service/history/historyEngine.go
@@ -41,7 +41,9 @@ import (
 	workflowpb "go.temporal.io/api/workflow/v1"
 	"go.temporal.io/api/workflowservice/v1"
 	sdkclient "go.temporal.io/sdk/client"
+
 	"go.temporal.io/server/common/persistence/visibility/manager"
+	"go.temporal.io/server/common/tasks"
 
 	enumsspb "go.temporal.io/server/api/enums/v1"
 	"go.temporal.io/server/api/historyservice/v1"
@@ -395,8 +397,8 @@ func (e *historyEngineImpl) registerNamespaceFailoverCallback() {
 				now := e.shard.GetTimeSource().Now()
 				// the fake tasks will not be actually used, we just need to make sure
 				// its length > 0 and has correct timestamp, to trigger a db scan
-				fakeWorkflowTask := []persistence.Task{&persistence.WorkflowTask{}}
-				fakeWorkflowTaskTimeoutTask := []persistence.Task{&persistence.WorkflowTaskTimeoutTask{VisibilityTimestamp: now}}
+				fakeWorkflowTask := []tasks.Task{&tasks.WorkflowTask{}}
+				fakeWorkflowTaskTimeoutTask := []tasks.Task{&tasks.WorkflowTaskTimeoutTask{VisibilityTimestamp: now}}
 				e.txProcessor.NotifyNewTask(e.currentClusterName, fakeWorkflowTask)
 				e.timerProcessor.NotifyNewTimers(e.currentClusterName, fakeWorkflowTaskTimeoutTask)
 			}
@@ -2261,8 +2263,8 @@ func (e *historyEngineImpl) SyncShardStatus(
 	// 2. notify the timer gate in the timer queue standby processor
 	// 3, notify the transfer (essentially a no op, just put it here so it looks symmetric)
 	e.shard.SetCurrentTime(clusterName, now)
-	e.txProcessor.NotifyNewTask(clusterName, []persistence.Task{})
-	e.timerProcessor.NotifyNewTimers(clusterName, []persistence.Task{})
+	e.txProcessor.NotifyNewTask(clusterName, []tasks.Task{})
+	e.timerProcessor.NotifyNewTimers(clusterName, []tasks.Task{})
 	return nil
 }
 
@@ -2548,7 +2550,7 @@ func (e *historyEngineImpl) NotifyNewHistoryEvent(
 }
 
 func (e *historyEngineImpl) NotifyNewTransferTasks(
-	tasks []persistence.Task,
+	tasks []tasks.Task,
 ) {
 
 	if len(tasks) > 0 {
@@ -2559,7 +2561,7 @@ func (e *historyEngineImpl) NotifyNewTransferTasks(
 }
 
 func (e *historyEngineImpl) NotifyNewTimerTasks(
-	tasks []persistence.Task,
+	tasks []tasks.Task,
 ) {
 
 	if len(tasks) > 0 {
@@ -2570,7 +2572,7 @@ func (e *historyEngineImpl) NotifyNewTimerTasks(
 }
 
 func (e *historyEngineImpl) NotifyNewReplicationTasks(
-	tasks []persistence.Task,
+	tasks []tasks.Task,
 ) {
 
 	if len(tasks) > 0 && e.replicatorProcessor != nil {
@@ -2579,7 +2581,7 @@ func (e *historyEngineImpl) NotifyNewReplicationTasks(
 }
 
 func (e *historyEngineImpl) NotifyNewVisibilityTasks(
-	tasks []persistence.Task,
+	tasks []tasks.Task,
 ) {
 
 	if len(tasks) > 0 && e.visibilityProcessor != nil {
