+++ service/history/historyEngineInterfaces.go
@@ -34,8 +34,8 @@ import (
 	persistencespb "go.temporal.io/server/api/persistence/v1"
 	replicationspb "go.temporal.io/server/api/replication/v1"
 	"go.temporal.io/server/common"
-	"go.temporal.io/server/common/persistence"
 	"go.temporal.io/server/common/task"
+	"go.temporal.io/server/common/tasks"
 	"go.temporal.io/server/service/history/shard"
 )
 
@@ -113,7 +113,7 @@ type (
 
 	timerProcessor interface {
 		taskExecutor
-		notifyNewTimers(timerTask []persistence.Task)
+		notifyNewTimers(timerTask []tasks.Task)
 	}
 
 	timerQueueAckMgr interface {
