+++ service/history/shard/engine_mock.go
@@ -37,7 +37,7 @@ import (
 	history "go.temporal.io/api/history/v1"
 	historyservice "go.temporal.io/server/api/historyservice/v1"
 	repication "go.temporal.io/server/api/replication/v1"
-	persistence "go.temporal.io/server/common/persistence"
+	tasks "go.temporal.io/server/common/tasks"
 	events "go.temporal.io/server/service/history/events"
 )
 
@@ -182,7 +182,7 @@ func (mr *MockEngineMockRecorder) NotifyNewHistoryEvent(event interface{}) *gomo
 }
 
 // NotifyNewReplicationTasks mocks base method.
-func (m *MockEngine) NotifyNewReplicationTasks(tasks []persistence.Task) {
+func (m *MockEngine) NotifyNewReplicationTasks(tasks []tasks.Task) {
 	m.ctrl.T.Helper()
 	m.ctrl.Call(m, "NotifyNewReplicationTasks", tasks)
 }
@@ -194,7 +194,7 @@ func (mr *MockEngineMockRecorder) NotifyNewReplicationTasks(tasks interface{}) *
 }
 
 // NotifyNewTimerTasks mocks base method.
-func (m *MockEngine) NotifyNewTimerTasks(tasks []persistence.Task) {
+func (m *MockEngine) NotifyNewTimerTasks(tasks []tasks.Task) {
 	m.ctrl.T.Helper()
 	m.ctrl.Call(m, "NotifyNewTimerTasks", tasks)
 }
@@ -206,7 +206,7 @@ func (mr *MockEngineMockRecorder) NotifyNewTimerTasks(tasks interface{}) *gomock
 }
 
 // NotifyNewTransferTasks mocks base method.
-func (m *MockEngine) NotifyNewTransferTasks(tasks []persistence.Task) {
+func (m *MockEngine) NotifyNewTransferTasks(tasks []tasks.Task) {
 	m.ctrl.T.Helper()
 	m.ctrl.Call(m, "NotifyNewTransferTasks", tasks)
 }
@@ -218,7 +218,7 @@ func (mr *MockEngineMockRecorder) NotifyNewTransferTasks(tasks interface{}) *gom
 }
 
 // NotifyNewVisibilityTasks mocks base method.
-func (m *MockEngine) NotifyNewVisibilityTasks(tasks []persistence.Task) {
+func (m *MockEngine) NotifyNewVisibilityTasks(tasks []tasks.Task) {
 	m.ctrl.T.Helper()
 	m.ctrl.Call(m, "NotifyNewVisibilityTasks", tasks)
 }
