+++ service/history/timerQueueActiveProcessor.go
@@ -37,6 +37,7 @@ import (
 	"go.temporal.io/server/common/log/tag"
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/persistence"
+	"go.temporal.io/server/common/tasks"
 	"go.temporal.io/server/service/history/shard"
 )
 
@@ -282,7 +283,7 @@ func (t *timerQueueActiveProcessorImpl) getReadLevel() timerKey {
 // NotifyNewTimers - Notify the processor about the new active timer events arrival.
 // This should be called each time new timer events arrives, otherwise timers maybe fired unexpected.
 func (t *timerQueueActiveProcessorImpl) notifyNewTimers(
-	timerTasks []persistence.Task,
+	timerTasks []tasks.Task,
 ) {
 	t.timerQueueProcessorBase.notifyNewTimers(timerTasks)
 }
