+++ service/history/timerQueueProcessor.go
@@ -41,6 +41,7 @@ import (
 	"go.temporal.io/server/common/log/tag"
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/persistence"
+	"go.temporal.io/server/common/tasks"
 	"go.temporal.io/server/common/xdc"
 	"go.temporal.io/server/service/history/configs"
 	"go.temporal.io/server/service/history/shard"
@@ -54,7 +55,7 @@ type (
 	timerQueueProcessor interface {
 		common.Daemon
 		FailoverNamespace(namespaceIDs map[string]struct{})
-		NotifyNewTimers(clusterName string, timerTask []persistence.Task)
+		NotifyNewTimers(clusterName string, timerTask []tasks.Task)
 		LockTaskProcessing()
 		UnlockTaskProcessing()
 	}
@@ -183,7 +184,7 @@ func (t *timerQueueProcessorImpl) Stop() {
 // This should be called each time new timer arrives, otherwise timers maybe fired unexpected.
 func (t *timerQueueProcessorImpl) NotifyNewTimers(
 	clusterName string,
-	timerTasks []persistence.Task,
+	timerTasks []tasks.Task,
 ) {
 
 	if clusterName == t.currentClusterName {
