+++ service/history/timerQueueStandbyProcessor.go
@@ -33,7 +33,7 @@ import (
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/log/tag"
 	"go.temporal.io/server/common/metrics"
-	"go.temporal.io/server/common/persistence"
+	"go.temporal.io/server/common/tasks"
 	"go.temporal.io/server/common/xdc"
 	"go.temporal.io/server/service/history/shard"
 )
@@ -186,7 +186,7 @@ func (t *timerQueueStandbyProcessorImpl) getReadLevel() timerKey {
 // NotifyNewTimers - Notify the processor about the new standby timer events arrival.
 // This should be called each time new timer events arrives, otherwise timers maybe fired unexpected.
 func (t *timerQueueStandbyProcessorImpl) notifyNewTimers(
-	timerTasks []persistence.Task,
+	timerTasks []tasks.Task,
 ) {
 
 	t.timerQueueProcessorBase.notifyNewTimers(timerTasks)
