+++ service/history/timerQueueTaskExecutorBase.go
@@ -38,6 +38,7 @@ import (
 	"go.temporal.io/server/common/namespace"
 	"go.temporal.io/server/common/persistence"
 	"go.temporal.io/server/common/searchattribute"
+	"go.temporal.io/server/common/tasks"
 	"go.temporal.io/server/service/history/configs"
 	"go.temporal.io/server/service/history/shard"
 	"go.temporal.io/server/service/history/workflow"
@@ -300,7 +301,7 @@ func (t *timerQueueTaskExecutorBase) deleteWorkflowVisibility(
 		TransferTasks:    nil,
 		TimerTasks:       nil,
 		ReplicationTasks: nil,
-		VisibilityTasks: []persistence.Task{&persistence.DeleteExecutionVisibilityTask{
+		VisibilityTasks: []tasks.Task{&tasks.DeleteExecutionVisibilityTask{
 			// TaskID is set by shard
 			VisibilityTimestamp: t.shard.GetTimeSource().Now(),
 			Version:             task.GetVersion(),
