+++ service/history/visibilityQueueProcessor.go
@@ -40,6 +40,7 @@ import (
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/persistence"
 	"go.temporal.io/server/common/persistence/visibility/manager"
+	"go.temporal.io/server/common/tasks"
 	"go.temporal.io/server/service/history/configs"
 	"go.temporal.io/server/service/history/shard"
 )
@@ -47,7 +48,7 @@ import (
 type (
 	visibilityQueueProcessor interface {
 		common.Daemon
-		NotifyNewTask(visibilityTasks []persistence.Task)
+		NotifyNewTask(visibilityTasks []tasks.Task)
 	}
 
 	updateVisibilityAckLevel func(ackLevel int64) error
@@ -213,7 +214,7 @@ func (t *visibilityQueueProcessorImpl) Stop() {
 // NotifyNewTask - Notify the processor about the new visibility task arrival.
 // This should be called each time new visibility task arrives, otherwise tasks maybe delayed.
 func (t *visibilityQueueProcessorImpl) NotifyNewTask(
-	visibilityTasks []persistence.Task,
+	visibilityTasks []tasks.Task,
 ) {
 	if len(visibilityTasks) != 0 {
 		t.notifyNewTask()
