+++ service/history/workflow/task_generator.go
@@ -38,8 +38,8 @@ import (
 	"go.temporal.io/server/common/clock"
 	"go.temporal.io/server/common/log"
 	"go.temporal.io/server/common/namespace"
-	"go.temporal.io/server/common/persistence"
 	"go.temporal.io/server/common/primitives/timestamp"
+	"go.temporal.io/server/common/tasks"
 )
 
 type (
@@ -143,7 +143,7 @@ func (r *TaskGeneratorImpl) GenerateWorkflowStartTasks(
 		return nil
 	}
 
-	r.mutableState.AddTimerTasks(&persistence.WorkflowTimeoutTask{
+	r.mutableState.AddTimerTasks(&tasks.WorkflowTimeoutTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: workflowRunExpirationTime,
 		Version:             startVersion,
@@ -159,13 +159,13 @@ func (r *TaskGeneratorImpl) GenerateWorkflowCloseTasks(
 	currentVersion := r.mutableState.GetCurrentVersion()
 	executionInfo := r.mutableState.GetExecutionInfo()
 
-	r.mutableState.AddTransferTasks(&persistence.CloseExecutionTask{
+	r.mutableState.AddTransferTasks(&tasks.CloseExecutionTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: now,
 		Version:             currentVersion,
 	})
 
-	r.mutableState.AddVisibilityTasks(&persistence.CloseExecutionVisibilityTask{
+	r.mutableState.AddVisibilityTasks(&tasks.CloseExecutionVisibilityTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: now,
 		Version:             currentVersion,
@@ -182,7 +182,7 @@ func (r *TaskGeneratorImpl) GenerateWorkflowCloseTasks(
 		return err
 	}
 
-	r.mutableState.AddTimerTasks(&persistence.DeleteHistoryEventTask{
+	r.mutableState.AddTimerTasks(&tasks.DeleteHistoryEventTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: now.Add(retention),
 		Version:             currentVersion,
@@ -212,7 +212,7 @@ func (r *TaskGeneratorImpl) GenerateDelayedWorkflowTasks(
 		return serviceerror.NewInternal(fmt.Sprintf("unknown initiator: %v", startAttr.GetInitiator()))
 	}
 
-	r.mutableState.AddTimerTasks(&persistence.WorkflowBackoffTimerTask{
+	r.mutableState.AddTimerTasks(&tasks.WorkflowBackoffTimerTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: executionTimestamp,
 		WorkflowBackoffType: workflowBackoffType,
@@ -229,7 +229,7 @@ func (r *TaskGeneratorImpl) GenerateRecordWorkflowStartedTasks(
 
 	startVersion := startEvent.GetVersion()
 
-	r.mutableState.AddVisibilityTasks(&persistence.StartExecutionVisibilityTask{
+	r.mutableState.AddVisibilityTasks(&tasks.StartExecutionVisibilityTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: now,
 		Version:             startVersion,
@@ -250,7 +250,7 @@ func (r *TaskGeneratorImpl) GenerateScheduleWorkflowTaskTasks(
 		return serviceerror.NewInternal(fmt.Sprintf("it could be a bug, cannot get pending workflow task: %v", workflowTaskScheduleID))
 	}
 
-	r.mutableState.AddTransferTasks(&persistence.WorkflowTask{
+	r.mutableState.AddTransferTasks(&tasks.WorkflowTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: now,
 		NamespaceID:         executionInfo.NamespaceId,
@@ -263,7 +263,7 @@ func (r *TaskGeneratorImpl) GenerateScheduleWorkflowTaskTasks(
 		scheduledTime := timestamp.TimeValue(workflowTask.ScheduledTime)
 		scheduleToStartTimeout := timestamp.DurationValue(r.mutableState.GetExecutionInfo().StickyScheduleToStartTimeout)
 
-		r.mutableState.AddTimerTasks(&persistence.WorkflowTaskTimeoutTask{
+		r.mutableState.AddTimerTasks(&tasks.WorkflowTaskTimeoutTask{
 			// TaskID is set by shard
 			VisibilityTimestamp: scheduledTime.Add(scheduleToStartTimeout),
 			TimeoutType:         enumspb.TIMEOUT_TYPE_SCHEDULE_TO_START,
@@ -291,7 +291,7 @@ func (r *TaskGeneratorImpl) GenerateStartWorkflowTaskTasks(
 	startedTime := timestamp.TimeValue(workflowTask.StartedTime)
 	workflowTaskTimeout := timestamp.DurationValue(workflowTask.WorkflowTaskTimeout)
 
-	r.mutableState.AddTimerTasks(&persistence.WorkflowTaskTimeoutTask{
+	r.mutableState.AddTimerTasks(&tasks.WorkflowTaskTimeoutTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: startedTime.Add(workflowTaskTimeout),
 		TimeoutType:         enumspb.TIMEOUT_TYPE_START_TO_CLOSE,
@@ -331,7 +331,7 @@ func (r *TaskGeneratorImpl) GenerateActivityTransferTasks(
 		}
 	}
 
-	r.mutableState.AddTransferTasks(&persistence.ActivityTask{
+	r.mutableState.AddTransferTasks(&tasks.ActivityTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: now,
 		NamespaceID:         targetNamespaceID,
@@ -352,7 +352,7 @@ func (r *TaskGeneratorImpl) GenerateActivityRetryTasks(
 		return serviceerror.NewInternal(fmt.Sprintf("it could be a bug, cannot get pending activity: %v", activityScheduleID))
 	}
 
-	r.mutableState.AddTimerTasks(&persistence.ActivityRetryTimerTask{
+	r.mutableState.AddTimerTasks(&tasks.ActivityRetryTimerTask{
 		// TaskID is set by shard
 		Version:             ai.Version,
 		VisibilityTimestamp: *ai.ScheduledTime,
@@ -381,7 +381,7 @@ func (r *TaskGeneratorImpl) GenerateChildWorkflowTasks(
 		return err
 	}
 
-	r.mutableState.AddTransferTasks(&persistence.StartChildExecutionTask{
+	r.mutableState.AddTransferTasks(&tasks.StartChildExecutionTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: now,
 		TargetNamespaceID:   targetNamespaceID,
@@ -416,7 +416,7 @@ func (r *TaskGeneratorImpl) GenerateRequestCancelExternalTasks(
 		return err
 	}
 
-	r.mutableState.AddTransferTasks(&persistence.CancelExecutionTask{
+	r.mutableState.AddTransferTasks(&tasks.CancelExecutionTask{
 		// TaskID is set by shard
 		VisibilityTimestamp:     now,
 		TargetNamespaceID:       targetNamespaceID,
@@ -453,7 +453,7 @@ func (r *TaskGeneratorImpl) GenerateSignalExternalTasks(
 		return err
 	}
 
-	r.mutableState.AddTransferTasks(&persistence.SignalExecutionTask{
+	r.mutableState.AddTransferTasks(&tasks.SignalExecutionTask{
 		// TaskID is set by shard
 		VisibilityTimestamp:     now,
 		TargetNamespaceID:       targetNamespaceID,
@@ -473,7 +473,7 @@ func (r *TaskGeneratorImpl) GenerateWorkflowSearchAttrTasks(
 
 	currentVersion := r.mutableState.GetCurrentVersion()
 
-	r.mutableState.AddVisibilityTasks(&persistence.UpsertExecutionVisibilityTask{
+	r.mutableState.AddVisibilityTasks(&tasks.UpsertExecutionVisibilityTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: now,
 		Version:             currentVersion, // task processing does not check this version
@@ -487,7 +487,7 @@ func (r *TaskGeneratorImpl) GenerateWorkflowResetTasks(
 
 	currentVersion := r.mutableState.GetCurrentVersion()
 
-	r.mutableState.AddTransferTasks(&persistence.ResetWorkflowTask{
+	r.mutableState.AddTransferTasks(&tasks.ResetWorkflowTask{
 		// TaskID is set by shard
 		VisibilityTimestamp: now,
 		Version:             currentVersion,
